Reworked tests for office-document.spec.ts to be in SubCategoryComponent.utils.spec.tsx
